// utils/editor/enums.ts
export enum ElementEnums {
  Paragraph = 'paragraph',
  HeadingOne = 'heading-one',
  HeadingTwo = 'heading-two',
  ListItem = 'list-item',
  NumberedList = 'numbered-list',
  BulletedList = 'bulleted-list',
  IFrame = 'iframe',
  Image = 'image',
  Link = 'link',
  File = 'file',
  Audio = 'audio',
}